fixed formatting issues (tabs -> spaces) added missing copyright headers moved camera script to development folder and fixed namespace implemented setLayer/setRenderer where missing reverted default scene to use default model added Rider specific gitignore entries added missing meta files to VCS
