+++ .gitignore
@@ -22,6 +22,10 @@
 # Visual Studio cache directory
 .vs/
 
+# Jetbrains Rider Cache
+.idea/
+Assets/Plugins/Editor/JetBrains*
+
 # Gradle cache directory
 .gradle/
