+++ Assets/Development/FreeFloatEditorCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c5717e119b4bff8408750aaea3ff68e0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
