+++ Assets/Example/LoadRexFileExample.unity
@@ -380,6 +380,5 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   rexFileNames:
-  - D&S_Foyer_P3DMKII_Modellkoordinatensystem.rex
-  - Deepspacebim.rex
+  - oscar.rex
   meshesToWrite: []
