+++ Assets/package/Runtime/Prefabs/PointListClusteredParticles.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c0c63be978eb9b5499a7788902997162
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
