+++ Assets/package/Runtime/Prefabs/PointListParticles.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3e7021d9e37b7ec418802fca5a6b8d97
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
