+++ Assets/package/Runtime/Prefabs/StandardPrefabClasses/PointListClusteredParticlesPrefab.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b63e1d30863ec34408566bd9eb36a95d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
