+++ Assets/package/Runtime/Prefabs/StandardPrefabClasses/PointListParticlesPrefab.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 56a8a9514b3145044865cffb2e0b93ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
