+++ Assets/package/Runtime/Prefabs/StandardPrefabClasses/ThreadedClustering.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 12ceb2b74ab57f74f8f0079f66e99fb0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
