+++ Packages/manifest.json
@@ -2,18 +2,18 @@
   "dependencies": {
     "com.unity.2d.sprite": "1.0.0",
     "com.unity.2d.tilemap": "1.0.0",
-    "com.unity.ads": "3.4.9",
+    "com.unity.ads": "3.6.1",
     "com.unity.analytics": "3.3.5",
     "com.unity.collab-proxy": "1.2.16",
     "com.unity.ide.rider": "1.1.4",
-    "com.unity.ide.vscode": "1.2.1",
-    "com.unity.multiplayer-hlapi": "1.0.6",
-    "com.unity.purchasing": "2.1.0",
-    "com.unity.test-framework": "1.1.16",
-    "com.unity.textmeshpro": "2.0.1",
-    "com.unity.timeline": "1.2.6",
+    "com.unity.ide.vscode": "1.2.3",
+    "com.unity.multiplayer-hlapi": "1.0.8",
+    "com.unity.purchasing": "3.0.1",
+    "com.unity.test-framework": "1.1.24",
+    "com.unity.textmeshpro": "2.1.4",
+    "com.unity.timeline": "1.2.18",
     "com.unity.ugui": "1.0.0",
-    "com.unity.xr.legacyinputhelpers": "2.1.4",
+    "com.unity.xr.legacyinputhelpers": "2.1.7",
     "com.unity.modules.ai": "1.0.0",
     "com.unity.modules.androidjni": "1.0.0",
     "com.unity.modules.animation": "1.0.0",
