+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.4.10f1
-m_EditorVersionWithRevision: 2019.4.10f1 (5311b3af6f69)
+m_EditorVersion: 2019.4.24f1
+m_EditorVersionWithRevision: 2019.4.24f1 (5da6f0345e82)
